Re-ran the grid combiner for the settlement froducts in the west.
To avoid overlap between the Michigan data I added a line to include the `cell` from which a point is obtained.  This allows us to accept the plss product in northern Michigan without re-writing, or duplicating it with the southern MI product.

Updated the version for the `full_paleon_conversion` to v0.3-2