Added Acknowledgements, fixed the reference file.
Fixed the figure order & output.

Fixed up references in the bibliography and related them in the paper.  Also soe minor parentheses issues with the references.  Now fixed.

Fixed a persistent issue with the transect plots.  num.rast was mis-assigned (with a different extent than expected) and so was re-assigned in the function.  That fixes everything.