Text changes to reflect conversation with D.Mladenoff (21/10/2015).  Clarifying spatial patterns of change is critical here.
A bit more discussion of how land use and historical vegetation interact to result in modern patterns of change in the Introduction.

Fixed some of the 'et al's so they're italicized.

Clarified exactly how the FIA estimates were made.

Minor text changes, as indicated by DJM.

Clarified the section on "distance to novelty", but kept largely as is.

Added a section to the results that breaks down changes regionally (by forest type).

Added Andria's paper to the references.

`fig11` - made sure we're not resampling with replacement.