`misc.functions` - Cleaned up code to fix `lint` problems.  Added the raster generation block in `get_cells`, this might slow down the function, but ensures that the raster numbering is standardized throughout.
`base_calculation` - Added the cell numbering from base.rast to `spec.table` explicitly, rather than taking the rounded values that come out of `morisita` that are generated from the means of points within the cell.  Added a `veil` diameter as well as the mean diameter.  Added a block of `assert` statements to the bottom of the file so that the knitting will fail if they fail.

`variables` is an output for my own use, just checking for variable re-use.  Honestly, this should be reduced as much as possible, but I don't know how much effort I want to put into it.